Introduction

Welcome to the documentation for CausalELM.

What is CausalELM

CausalELM is a project to learn about how machine learning methods can be used for causal inference, gain a better understanding of Extreme Learning Machines, and build proficiency with Julia. The intutition for using ELM for causal inference is that in event studies and similar research designs, one must make a tradeoff between accuracy vs complexity and computing power when deciding on a model to estimate the counterfactual for the post-event/post-treatment period. ELMs represent a good balance because they have been shown to be very accurate while not having to rely on backpropagation for training. Furthermore, since ELMs solve the least squares problem, their estimates (with the design matrix that has been transformed by an activation function) represent the best linear unbiased estimate and are asymptotically consistent. While this is a research project, it is functional and has passed a large number of unit tests, so it should be mostly reliable. As the goal is to be able to use this package for causal inference I have added a couple structs and methods for that goal. First, the predictcounterfactual method uses a trained ExtremeLearningMachine struct to estimate the counterfactual, or post-event observations, using data that can be passed in by the user. Second, there is a method to perform a placebo test whereby counterfactuals are predicted using both pre-event and post-event covariates. If these predictions are significantly different there is likely a confounder or some other issue with the study design. Finally, to address the issue of multicollinearity the RegularizedExtremeLearner struct and associated methods fit and predict ELMs with L2 regularization, as proposed by Li and Niu (2013). My intent for this package is for it to be lightweight, so I wrote it such that it does not have any dependencies.

License

CausalELM is licensed under an MIT license!

CausalELM.ActivationFunctions.binarystep — Method

binarystep(x)

Apply the binary step activation function to an array or real number.

Examples

julia> binarystep(1)
1
julia> binarystep([-1000, 100, 1, 0, -0.001, -3])
[0, 1, 1, 1, 0, 0]

CausalELM.ActivationFunctions.elish — Method

elish(x)

Apply the ELiSH activation function to an array or real number.

Examples

julia> elish(1)
0.7310585786300049
julia> gaussian([-1, 1])
[-0.1700034015685479, 0.7310585786300049]

CausalELM.ActivationFunctions.fourier — Method

elish(x)

Apply the Fourier activation function to an array or real number.

Examples

julia> fourier(1)
0.8414709848078965
julia> fourier([-1, 1])
[-0.8414709848078965, 0.8414709848078965]

CausalELM.ActivationFunctions.gaussian — Method

gaussian(x)

Apply the gaussian activation function to an array or real number.

Examples

julia> gaussian(1)
0.11443511435028261
julia> gaussian([1, -1])
[0.36787944117144233, 0.36787944117144233]

CausalELM.ActivationFunctions.gelu — Method

gelu(x)

Apply the GeLU activation function to an array or real number.

Examples

julia> gelu(1)
0.8411919906082768
julia> gelu([-1, 0, 1])
[-0.15880800939172324, 0, 0.8411919906082768]

CausalELM.ActivationFunctions.hardtanh — Method

hardtanh(x)

Apply the hardtanh activation function to an array or real number.

Examples

julia> hardtanh(-2)
-1
julia> gaussian([-2, 0, 2])
[-1, 0, 1]

CausalELM.ActivationFunctions.leakyrelu — Method

relu(x)

Apply the leaky ReLU activation function to an array or real number.

Examples

julia> relu(1)
1
julia> relu([-0.01, 0, 1])
[1, 0, 0]

CausalELM.ActivationFunctions.relu — Method

relu(x)

Apply the ReLU activation function to an array or real number.

Examples

julia> relu(1)
1
julia> relu([1, 0, -1])
[1, 0, 0]

CausalELM.ActivationFunctions.softmax — Method

swish(x)

Apply the softmax activation function to an array or real number.

For numbers that have large absolute values this function might be numerically unstable.

Examples

julia> softmax(1)
2.718281828459045
julia> softmax([1, -1])
[2.718281828459045, -0.36787944117144233]

CausalELM.ActivationFunctions.softplus — Method

softplus(x)

Apply the softplus activation function to an array or real number.

For numbers that have large absolute values this function might be numerically unstable.

Examples

julia> softplus(1)
1.3132616875182228
julia> softplus([1, -1])
[1.3132616875182228, 0.31326168751822286]

CausalELM.ActivationFunctions.swish — Method

swish(x)

Apply the swish activation function to an array or real number.

Examples

julia> swish(1)
0.7310585786300049
julia> swish([1, 0, -1])
[0.7310585786300049, 0, -0.2689414213699951]

CausalELM.ActivationFunctions.tanh — Method

tanh(x)

Apply the tanh activation function to an array.

This is just a vectorized version of Base.tanh

Examples

julia> tanh([1, 0])
[0.7615941559557649, 0.0]

CausalELM.ActivationFunctions.σ — Method

σ(x)

Apply the sigmoid activation function to an array or real number.

Examples

julia> σ(1)
0.7310585786300049
julia> σ([1, 0])
[0.7310585786300049, 0.5]

CausalELM.Models.fit! — Method

fit!(model, activation)

Make predictions with an ExtremeLearner.

For more details see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Examples ```julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden nodes julia> f1 = fit!(m1, σ) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233]

CausalELM.Models.fit! — Method

fit!(model, activation)

Fit a Regularized Extreme Learner.

For more details see: Li, Guoqiang, and Peifeng Niu. "An enhanced extreme learning machine based on ridge regression for regression." Neural Computing and Applications 22, no. 3 (2013): 803-810.

Examples ```julia-repl julia> m1 = RegularizedExtremeLearner(x, y, 10, σ) Regularized Extreme Learning Machine with 10 hidden nodes julia> f1 = fit!(m1, σ) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233]

CausalELM.Models.placebotest — Method

placebotest(model)

Conduct a placebo test.

This method makes predictions for the post-event or post-treatment period using data in the pre-event or pre-treatment period and the post-event or post-treament. If there is a statistically significant difference between these predictions the study design may be flawed. Due to the multitude of significance tests for time series data, this function returns the predictions but does not test for statistical significance.

Examples ```julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden nodes julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predictcounterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978] julia> placebotest(m1) ([9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978], [0.5, 0.4, 0.3, 0.2])

CausalELM.Models.predict — Method

predict(model, X)

Use an ExtremeLearningMachine to make predictions.

For more details see: Huang G-B, Zhu Q-Y, Siew C. Extreme learning machine: theory and applications. Neurocomputing. 2006;70:489–501. https://doi.org/10.1016/j.neucom.2005.12.126

Examples ```julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden nodes julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predict(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978]

CausalELM.Models.predictcounterfactual! — Method

predictcounterfactual(model, X)

Use an ExtremeLearningMachine to predict the counterfactual.

This should be run with the observed covariates. To use synthtic data for what-if scenarios use predict.

See also 'predict'.

Examples ```julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden nodes julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predictcounterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978]

CausalELM.Models.ExtremeLearner — Type

ExtremeLearner(X, Y, hidden_nodes, activation)

Construct an ExtremeLearner for fitting and prediction.

While it is possible to use an ExtremeLearner for regression, it is recommended to use RegularizedExtremeLearner, which imposes an L2 penalty, to reduce multicollinearity.

For more details see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Examples ```julia-repl julia> x = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0] 4×2 Matrix{Float64}: 1.0 1.0 0.0 1.0 0.0 0.0 1.0 0.0 julia> y = [0.0, 1.0, 0.0, 1.0] 4-element Vector{Int64}: 0.0 1.0 0.0 1.0 julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden nodes

CausalELM.Models.RegularizedExtremeLearner — Type

RegularizedExtremeLearner(X, Y, hidden_nodes, activation)

Construct a RegularizedExtremeLearner for fitting and prediction.

For more details see: Li, Guoqiang, and Peifeng Niu. "An enhanced extreme learning machine based on ridge regression for regression." Neural Computing and Applications 22, no. 3 (2013): 803-810.

Examples ```julia-repl julia> x = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0] 4×2 Matrix{Float64}: 1.0 1.0 0.0 1.0 0.0 0.0 1.0 0.0 julia> y = [0.0, 1.0, 0.0, 1.0] 4-element Vector{Int64}: 0.0 1.0 0.0 1.0 julia> m1 = RegularizedExtremeLearner(x, y, 10, σ) Regularized Extreme Learning Machine with 10 hidden nodes