API

This is a reference for advanced users or those who wish to contribute.

CausalELM

CausalELM.ActivationFunctions

CausalELM.CrossValidation

CausalELM.Estimators

CausalELM.Inference

CausalELM.Metalearners

CausalELM.Metrics

CausalELM.Models

CausalELM.Estimators.CausalEstimator

CausalELM.Estimators.DoublyRobust

CausalELM.Estimators.EventStudy

CausalELM.Estimators.GComputation

CausalELM.Metalearners.Metalearner

CausalELM.Metalearners.SLearner

CausalELM.Metalearners.TLearner

CausalELM.Metalearners.XLearner

CausalELM.Models.ExtremeLearner

CausalELM.Models.ExtremeLearningMachine

CausalELM.Models.RegularizedExtremeLearner

CausalELM.ActivationFunctions.binarystep

CausalELM.ActivationFunctions.elish

CausalELM.ActivationFunctions.fourier

CausalELM.ActivationFunctions.gaussian

CausalELM.ActivationFunctions.gelu

CausalELM.ActivationFunctions.hardtanh

CausalELM.ActivationFunctions.leakyrelu

CausalELM.ActivationFunctions.relu

CausalELM.ActivationFunctions.softmax

CausalELM.ActivationFunctions.softplus

CausalELM.ActivationFunctions.swish

CausalELM.ActivationFunctions.tanh

CausalELM.ActivationFunctions.σ

CausalELM.CrossValidation.bestsize

CausalELM.CrossValidation.crossvalidate

CausalELM.CrossValidation.recode

CausalELM.CrossValidation.traintest

CausalELM.CrossValidation.validate

CausalELM.Inference.generatenulldistribution

CausalELM.Inference.quantitiesofinterest

CausalELM.Metrics.F1

CausalELM.Metrics.accuracy

CausalELM.Metrics.confusionmatrix

CausalELM.Metrics.mae

CausalELM.Metrics.mse

CausalELM.Metrics.precision

CausalELM.Metrics.recall

CausalELM.Models.fit!

CausalELM.Models.placebotest

CausalELM.Models.predict

CausalELM.Models.predictcounterfactual!

CausalELM.estimatecausaleffect!

CausalELM.summarize

CausalELM

CausalELM — Module

Macros, functions, and structs for applying Extreme Learning Machines to causal inference tasks where the counterfactual is unavailable or biased and must be predicted. Provides macros for event study designs, parametric G-computation, doubly robust machine learning, and metalearners. Additionally, these tasks can be performed with or without L2 penalization and will automatically choose the best number of neurons and L2 penalty.

For more details on Extreme Learning Machines see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Activation Functions

CausalELM.ActivationFunctions — Module

Activation functions for Extreme Learning machines

CausalELM.ActivationFunctions.binarystep — Function

binarystep(x)

Apply the binary step activation function to a real number.

Examples

julia> binarystep(1)
1

binarystep(x)

Apply the binary step activation function to an array.

Examples

julia> binarystep([-1000, 100, 1, 0, -0.001, -3])
[0, 1, 1, 1, 0, 0]

CausalELM.ActivationFunctions.σ — Function

σ(x)

Apply the sigmoid activation function to a real number.

Examples

julia> σ(1)
0.7310585786300049

σ(x)

Apply the sigmoid activation function to an array.

Examples

julia> σ([1, 0])
[0.7310585786300049, 0.5]

CausalELM.ActivationFunctions.tanh — Function

tanh(x)

Apply the tanh activation function to an array.

This is just a vectorized version of Base.tanh

Examples

julia> tanh([1, 0])
[0.7615941559557649, 0.0]

CausalELM.ActivationFunctions.relu — Function

relu(x)

Apply the ReLU activation function to a real number.

Examples

julia> relu(1)
1

relu(x)

Apply the ReLU activation function to an array.

Examples

julia> relu([1, 0, -1])
[1, 0, 0]

CausalELM.ActivationFunctions.leakyrelu — Function

leakyrelu(x)

Apply the leaky ReLU activation function to a real number.

Examples

julia> leakyrelu(1)
1

leakyrelu(x)

Apply the leaky ReLU activation function to an array.

Examples

julia> leakyrelu([-0.01, 0, 1])
[1, 0, 0]

CausalELM.ActivationFunctions.swish — Function

swish(x)

Apply the swish activation function to a real number.

Examples

julia> swish(1)
0.7310585786300049

swish(x)

Apply the swish activation function to an array.

Examples

julia> swish([1, 0, -1])
[0.7310585786300049, 0, -0.2689414213699951]

CausalELM.ActivationFunctions.softmax — Function

softmax(x)

Apply the softmax activation function to a real number.

For numbers that have large absolute values this function may become numerically unstable.

Examples

julia> softmax(1)
2.718281828459045

softmax(x)

Apply the softmax activation function to an array.

For numbers that have large absolute values this function might be numerically unstable.

Examples

julia> softmax([1, -1])
[2.718281828459045, -0.36787944117144233]

CausalELM.ActivationFunctions.softplus — Function

softplus(x)

Apply the softplus activation function to a real number.

Examples

julia> softplus(1)
1.3132616875182228

softplus(x)

Apply the softplus activation function to an array.

Examples

julia> softplus([1, -1])
[1.3132616875182228, 0.31326168751822286]

CausalELM.ActivationFunctions.gelu — Function

gelu(x)

Apply the GeLU activation function to a real number.

Examples

julia> gelu(1)
0.8411919906082768

gelu(x)

Apply the GeLU activation function to an array.

Examples

julia> gelu([-1, 0, 1])
[-0.15880800939172324, 0, 0.8411919906082768]

CausalELM.ActivationFunctions.gaussian — Function

gaussian(x)

Apply the gaussian activation function to a real number.

Examples

julia> gaussian(1)
0.11443511435028261

gaussian(x)

Apply the gaussian activation function to an array.

Examples

julia> gaussian([1, -1])
[0.36787944117144233, 0.36787944117144233]

CausalELM.ActivationFunctions.hardtanh — Function

hardtanh(x)

Apply the hardtanh activation function to a real number.

Examples

julia> hardtanh(-2)
-1

hardtanh(x)

Apply the hardtanh activation function to an array.

Examples

julia> hardtanh([-2, 0, 2])
[-1, 0, 1]

CausalELM.ActivationFunctions.elish — Function

elish(x)

Apply the ELiSH activation function to a real number.

Examples

julia> elish(1)
0.7310585786300049

elish(x)

Apply the ELiSH activation function to an array.

Examples

julia> elish([-1, 1])
[-0.17000340156854793, 0.7310585786300049]

CausalELM.ActivationFunctions.fourier — Function

fourrier(x)

Apply the Fourier activation function to a real number.

Examples

julia> fourier(1)
0.8414709848078965

fourrier(x)

Apply the Fourier activation function to an array.

Examples

julia> fourier([-1, 1])
[-0.8414709848078965, 0.8414709848078965]

Cross Valdiation

CausalELM.CrossValidation — Module

Methods to perform cross validation and find the optimum number of neurons.

To reduce computation time, the number of neurons is optimized by using cross validation to estimate the validation error on a small subset of the range of possible numbers of neurons. Then, an Extreme Learning Machine is trained to predict validation loss from the given cross validation sets. Finally, the number of neurons is selected that has the smallest predicted loss or the highest classification metric.

CausalELM.CrossValidation.recode — Function

recode(ŷ)

Round predicted values to their predicted class for classification tasks.

If the smallest predicted label is 0, all labels are shifted up 1; if the smallest label is -1, all labels are shifted up 2. Also labels cannot be smaller than -1.

Examples

julia> recode([-0.7, 0.2, 1.1])
3-element Vector{Float64}
1
2
3
julia> recode([0.1, 0.2, 0.3])
3-element Vector{Float64}
1
1
1
julia> recode([1.1, 1.51, 1.8])
3-element Vector{Float64}
1
2
2

CausalELM.CrossValidation.traintest — Function

traintest(X, Y, folds)

Create a train-test split.

If an iteration is specified, the train test split will be treated as time series/panel data.

Examples

julia> xtrain, ytrain, xtest, ytest = traintest(zeros(20, 2), zeros(20), 5)

traintest(X, Y, folds, iteration)

Create a rolling train-test split for time series/panel data.

An iteration should not be specified for non-time series/panel data.

Examples

julia> xtrain, ytrain, xtest, ytest = traintest(zeros(20, 2), zeros(20), 5, 1)

CausalELM.CrossValidation.validate — Function

validate(X, Y, nodes, metric, iteration...; activation, regularized, folds)

Calculate a validation metric for a single fold in k-fold cross validation.

Examples

julia> x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
julia> validate(x, y, 5, accuracy, 3)
0.0

CausalELM.CrossValidation.crossvalidate — Function

crossvalidate(X, Y, neurons, metric, activation, regularized, folds)

Calculate a validation metric for k folds using a single set of hyperparameters.

Examples

julia> x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
julia> crossvalidate(x, y, 5, accuracy)
0.0257841765251021

CausalELM.CrossValidation.bestsize — Function

bestsize(X, Y, metric, task, activation, min_neurons, max_neurons, regularized, folds, temporal, 
    iterations, approximator_neurons)

Compute the best number of neurons for an Extreme Learning Machine.

The procedure tests networks with numbers of neurons in a sequence whose length is given by iterations on the interval [minneurons, maxneurons]. Then, it uses the networks sizes and validation errors from the sequence to predict the validation error or metric for every network size between minneurons and maxneurons using the function approximation ability of an Extreme Learning Machine. Finally, it returns the network size with the best predicted validation error or metric.

Examples

julia> bestsize(rand(100, 5), rand(100), mse, "regression")
11

ATE/ATE/ITT Estimation

CausalELM.Estimators — Module

Estimate causal effects with event study designs, G-computation, and doubly robust estiamtion using Extreme Learning machines.

CausalELM.Estimators.CausalEstimator — Type

Abstract type for GComputation and DoublyRobust

CausalELM.Estimators.EventStudy — Type

Container for the results of an event study

CausalELM.Estimators.GComputation — Type

Container for the results of G-Computation

CausalELM.Estimators.DoublyRobust — Type

Container for the results of doubly robust estimation

CausalELM.estimatecausaleffect! — Function

estimatecausaleffect!(study)

Estimate the abnormal returns in an event study.

Examples

julia> X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
julia> m1 = EventStudy(X₀, Y₀, X₁, Y₁)
julia> estimatecausaleffect!(m1)
0.25714308

estimatecausaleffect!(g)

Estimate a causal effect of interest using G-Computation.

If treatents are administered at multiple time periods, the effect will be estimated as the average difference between the outcome of being treated in all periods and being treated in no periods. For example, given that individuals 1, 2, ..., i ∈ I recieved either a treatment or a placebo in p different periods, the model would estimate the average treatment effect as E[Yᵢ|T₁=1, T₂=1, ... Tₚ=1, Xₚ] - E[Yᵢ|T₁=0, T₂=0, ... Tₚ=0, Xₚ].

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = GComputation(X, Y, T)
julia> estimatecausaleffect!(m1)
0.31067439

estimatecausaleffect!(DRE)

Estimate a causal effect of interest using doubly robust estimation.

Unlike other estimators, this method does not support time series or panel data. This method also does not work as well with smaller datasets because it estimates separate outcome models for the treatment and control groups.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = DoublyRobust(X, Y, T)
julia> estimatecausaleffect!(m1)
0.31067439

estimatecausaleffect!(s)

Estimate the CATE using an S-Learner.

For an overview of meatlearning, including S-Learners see:

Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for 
estimating heterogeneous treatment effects using machine learning." Proceedings of the 
national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = SLearner(X, Y, T)
julia> estimatecausaleffect!(m1)
[0.20729633391630697, 0.20729633391630697, 0.20729633391630692, 0.20729633391630697, 
0.20729633391630697, 0.20729633391630697, 0.20729633391630697, 0.20729633391630703, 
0.20729633391630697, 0.20729633391630697  …  0.20729633391630703, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630703, 0.20729633391630697, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630697, 0.20729633391630697, 0.20729633391630697]

estimatecausaleffect!(t)

Estimate the CATE using a T-Learner.

For an overview of meatlearning, including T-Learners see:

Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for 
estimating heterogeneous treatment effects using machine learning." Proceedings of the 
national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = TLearner(X, Y, T)
julia> estimatecausaleffect!(m1)
[0.0493951571746305, 0.049395157174630444, 0.0493951571746305, 0.049395157174630444, 
0.04939515717463039, 0.04939515717463039, 0.04939515717463039, 0.04939515717463039, 
0.049395157174630444, 0.04939515717463061  …  0.0493951571746305, 0.04939515717463039, 
0.0493951571746305, 0.04939515717463039, 0.0493951571746305, 0.04939515717463039, 
0.04939515717463039, 0.049395157174630444, 0.04939515717463039, 0.049395157174630444]

estimatecausaleffect!(x)

Estimate the CATE using an X-Learner.

For an overview of meatlearning, including X-Learners see:

Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for 
estimating heterogeneous treatment effects using machine learning." Proceedings of the 
national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = XLearner(X, Y, T)
julia> estimatecausaleffect!(m1)
[-0.025012644892878473, -0.024634294305967294, -0.022144246680543364, -0.023983138957276127, 
-0.024756239357838557, -0.019409519377053822, -0.02312807640357356, -0.016967113188439076, 
-0.020188871831409317, -0.02546526148141366  …  -0.019811641136866287, 
-0.020780821058711863, -0.013588359417922776, -0.020438648396328824, -0.016169487825519843, 
-0.024031422484491572, -0.01884713946778991, -0.021163590874553318, -0.014607310062509895, 
-0.022449034332142046]

CATE Estimation

CausalELM.Metalearners — Module

Metalearners to estimate the conditional average treatment effect (CATE).

CausalELM.Metalearners.Metalearner — Type

Abstract type for metalearners

CausalELM.Metalearners.SLearner — Type

S-Learner for CATE estimation.

CausalELM.Metalearners.TLearner — Type

T-Learner for CATE estimation.

CausalELM.Metalearners.XLearner — Type

X-Learner for CATE estimation.

Missing docstring.

Missing docstring for CausalELM.Metalearners.SLearner.estimatecausaleffect!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CausalELM.Metalearners.TLearner.estimatecausaleffect!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CausalELM.Metalearners.XLearner.estimatecausaleffect!. Check Documenter's build log for details.

Inference and Summarization

CausalELM.Inference — Module

Methods for summarization and inference from estimators and metalearners.

CausalELM.summarize — Function

summarize(study, mean_effect)

Return a summary from an event study.

Examples

julia> X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
julia> m1 = EventStudy(X₀, Y₀, X₁, Y₁)
julia> estimatetreatmenteffect!(m1)
[0.25714308]
julia> summarize(m1)
{"Task" => "Regression", "Regularized" => true, "Activation Function" => relu, 
"Validation Metric" => "mse","Number of Neurons" => 2, 
"Number of Neurons in Approximator" => 10, "β" => [0.25714308], 
"Causal Effect" => -3.9101138, "Standard Error" => 1.903434356, "p-value" = 0.00123356}

summarize(g)

Return a summary from a G-Computation estimator.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = GComputation(X, Y, T)
julia> estimatetreatmenteffect!(m1)
[0.3100468253]
julia> summarize(m1)
{"Task" => "Regression", "Quantity of Interest" => "ATE", Regularized" => "true", 
"Activation Function" => "relu", "Time Series/Panel Data" => "false", 
"Validation Metric" => "mse","Number of Neurons" => "5", 
"Number of Neurons in Approximator" => "10", "β" => "[0.3100468253]",
"Causal Effect: 0.00589761, "Standard Error" => 5.12900734, "p-value" => 0.479011245} 

summarize(dre, n)

Return a summary from a doubly robust estimator.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = DoublyRobust(X, X, Y, T)
julia> estimatetreatmenteffect!(m1)
[0.5804032956]
julia> summarize(m1)
{"Task" => "Regression", "Quantity of Interest" => "ATE", Regularized" => "true", 
"Activation Function" => "relu", "Validation Metric" => "mse", "Number of Neurons" => "5", 
"Number of Neurons in Approximator" => "10", "Causal Effect" = 0.5804032956, 
"Standard Error" => 2.129400324, "p-value" => 0.0008342356}

summarize(m, n)

Return a summary from a metalearner.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = SLearner(X, Y, T)
julia> estimatecate!(m1)
[0.20729633391630697, 0.20729633391630697, 0.20729633391630692, 0.20729633391630697, 
0.20729633391630697, 0.20729633391630697, 0.20729633391630697, 0.20729633391630703, 
0.20729633391630697, 0.20729633391630697  …  0.20729633391630703, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630703, 0.20729633391630697, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630697, 0.20729633391630697, 0.20729633391630697]
julia> summarise(m1)
{"Task" => "Regression", Regularized" => "true", "Activation Function" => "relu", 
"Time Series/Panel Data" => "false", "Validation Metric" => "mse", 
"Number of Neurons" => "5", "Number of Neurons in Approximator" => "10", 
"β" => "[0.3100468253]", "Causal Effect: [0.20729633391630697, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630697, 0.20729633391630697, 0.20729633391630697, 
0.20729633391630697, 0.20729633391630703, 0.20729633391630697, 0.20729633391630697  …  
0.20729633391630703, 0.20729633391630697, 0.20729633391630692, 0.20729633391630703, 
0.20729633391630697, 0.20729633391630697, 0.20729633391630692, 0.20729633391630697, 
0.20729633391630697, 0.20729633391630697], "Standard Error" => 5.3121435085, 
"p-value" => 0.0632454855}

CausalELM.Inference.quantitiesofinterest — Function

quantitiesofinterest(model, n)

Generate a p-value and standard error through randomization inference

This method generates a null distribution of treatment effects by reestimating treatment effects from permutations of the treatment vector and estimates a p-value and standard from the generated distribution.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

julia> x, y, t = rand(100, 5), rand(1:100, 100, 1), [rand()<0.4 for i in 1:100]
julia> g_computer = GComputation(x, y, t)
julia> estimatecausaleffect!(g_computer)
julia> quantitiesofinterest(g_computer, 1000)
(0.114, 6.953133617011371)

quantitiesofinterest(model, nsplits)

Generate a p-value and standard error through randomization inference

This method generates a null distribution of treatment effects by reestimating treatment effects from permutations of the treatment vector and estimates a p-value and standard from the generated distribution. Randomization for event studies is done by creating time splits at even intervals and reestimating the causal effect.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

julia> x₀, y₀, x₁, y₁ = rand(1:100, 100, 5), rand(100), rand(10, 5), rand(10)
julia> event_study = EventStudy(x₀, y₀, x₁, y₁)
julia> estimatecausaleffect!(event_study)
julia> quantitiesofinterest(event_study, 10)
(0.0, 0.07703275541001667)

CausalELM.Inference.generatenulldistribution — Function

generatenulldistribution(e, n)

Generate a null distribution for the treatment effect of G-computation, doubly robust estimation, or metalearning.

This method estimates the same model that is provided using random permutations of the treatment assignment to generate a vector of estimated effects under different treatment regimes. When e is a metalearner the null statistic is the difference is the ATE.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

Examples

julia> x, y, t = rand(100, 5), rand(1:100, 100, 1), [rand()<0.4 for i in 1:100]
julia> g_computer = GComputation(x, y, t)
julia> estimatecausaleffect!(g_computer)
julia> generatenulldistribution(g_computer, 500)
[0.016297180690693656, 0.0635928694685571, 0.20004144093635673, 0.505893866040335, 
0.5130594630907543, 0.5432486130493388, 0.6181727442724846, 0.61838399963459, 
0.7038981488009489, 0.7043407710415689  …  21.909186142780246, 21.960498059428854, 
21.988553083790023, 22.285403459215363, 22.613625375395973, 23.382102081355548, 
23.52056245175936, 24.739658523175912, 25.30523686137909, 28.07474553316176]

generatenulldistribution(e, n, mean_effect)

Generate a null distribution for the treatment effect in an event study design. By default, this method generates a null distribution of mean differences. To generate a null distribution of cummulative differences, set the mean_effect argument to false.

Instead of randomizing the assignment of units to the treamtent or control group, this method generates the null distribution by reestimating the event study with the intervention set to n splits at even intervals within the total study duration.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

julia> x₀, y₀, x₁, y₁ = rand(1:100, 100, 5), rand(100), rand(10, 5), rand(10)
julia> event_study = EventStudy(x₀, y₀, x₁, y₁)
julia> estimatecausaleffect!(event_study)
julia> generatenulldistribution(event_study, 10)
[-0.5012456678829079, -0.33790650529972194, -0.2534340182760628, -0.21030239864895905, 
-0.11672915615117885, -0.08149441936166794, -0.0685134758182695, -0.06217013151235991, 
-0.05905529159312335, -0.04927743270606937]

Validation Metrics

CausalELM.Metrics — Module

Metrics to evaluate the performance of an Extreme learning machine for regression and classification tasks.

CausalELM.Metrics.mse — Function

mse(y, ŷ)

Calculate the mean squared error

See also mae.

Examples

julia> mse([0.0, 0.0, 0.0], [0.0, 0.0, 0.0])
0
julia> mse([-1.0, -1.0, -1.0], [1.0, 1.0, 1.0])
4

CausalELM.Metrics.mae — Function

mae(y, ŷ)

Calculate the mean absolute error

See also mse.

Examples

julia> mae([-1.0, -1.0, -1.0], [1.0, 1.0, 1.0])
2
julia> mae([1.0, 1.0, 1.0], [2.0, 2.0, 2.0])
1

CausalELM.Metrics.confusionmatrix — Function

confusionmatrix(y, ŷ)

Generate a confusion matrix

Examples

julia> confusionmatrix([1, 1, 1, 1, 0], [1, 1, 1, 1, 0])
2×2 Matrix{Int64}:
 1  0
 0 4
julia> confusionmatrix([1, 1, 1, 1, 0, 2], [1, 1, 1, 1, 0, 2])
3×3 Matrix{Int64}:
 1  0 0
 0 4 0
 0 0 1

CausalELM.Metrics.accuracy — Function

accuracy(y, ŷ)

Calculate the accuracy for a classification task

Examples

julia> accuracy([1, 1, 1, 1], [0, 1, 1, 0])
0.5
julia> accuracy([1, 2, 3, 4], [1, 1, 1, 1])
0.25

CausalELM.Metrics.precision — Function

precision(y, ŷ)

Calculate the precision for a classification task

See also recall.

Examples

julia> precision([0, 1, 0, 0], [0, 1, 1, 0])
0.5
julia> precision([0, 1, 0, 0], [0, 1, 0, 0])
1

CausalELM.Metrics.recall — Function

recall(y, ŷ)

Calculate the recall for a classification task

See also precision.

Examples

julia> recall([1, 2, 1, 3, 0], [2, 2, 2, 3, 1])
0.5
julia> recall([1, 2, 1, 3, 2], [2, 2, 2, 3, 1])
1

CausalELM.Metrics.F1 — Function

F1(y, ŷ)

Calculate the F1 score for a classification task

Examples

julia> F1([1, 2, 1, 3, 0], [2, 2, 2, 3, 1])
0.4
julia> F1([1, 2, 1, 3, 2], [2, 2, 2, 3, 1])
0.47058823529411764

Base Models

CausalELM.Models — Module

Base models to perform extreme learning with and without L2 penalization.

For details on Extreme learning machines see; Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

For details on Extreme learning machines with an L2 penalty see: Li, Guoqiang, and Peifeng Niu. "An enhanced extreme learning machine based on ridge regression for regression." Neural Computing and Applications 22, no. 3 (2013): 803-810.

CausalELM.Models.ExtremeLearningMachine — Type

Abstract type that includes vanilla and L2 regularized Extreme Learning Machines

CausalELM.Models.ExtremeLearner — Type

Struct to hold data for an Extreme Learning machine

CausalELM.Models.RegularizedExtremeLearner — Type

Struct to hold data for a regularized Extreme Learning Machine

CausalELM.Models.fit! — Function

fit!(model)

Make predictions with an ExtremeLearner.

For more details see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Examples julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden neurons julia> f1 = fit!(m1) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233]

fit!(model)

Fit a Regularized Extreme Learner.

For more details see: Li, Guoqiang, and Peifeng Niu. "An enhanced extreme learning machine based on ridge regression for regression." Neural Computing and Applications 22, no. 3 (2013): 803-810.

Examples julia-repl julia> m1 = RegularizedExtremeLearner(x, y, 10, σ) Regularized Extreme Learning Machine with 10 hidden neurons julia> f1 = fit!(m1) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233]

CausalELM.Models.predict — Function

predict(model, X)

Use an ExtremeLearningMachine to make predictions.

For more details see: Huang G-B, Zhu Q-Y, Siew C. Extreme learning machine: theory and applications. Neurocomputing. 2006;70:489–501. https://doi.org/10.1016/j.neucom.2005.12.126

Examples julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden neurons julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predict(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978]

CausalELM.Models.predictcounterfactual! — Function

predictcounterfactual(model, X)

Use an ExtremeLearningMachine to predict the counterfactual.

This should be run with the observed covariates. To use synthtic data for what-if scenarios use predict.

See also predict.

Examples julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden neurons julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predictcounterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978]

CausalELM.Models.placebotest — Function

placebotest(model)

Conduct a placebo test.

This method makes predictions for the post-event or post-treatment period using data in the pre-event or pre-treatment period and the post-event or post-treament. If there is a statistically significant difference between these predictions the study design may be flawed. Due to the multitude of significance tests for time series data, this function returns the predictions but does not test for statistical significance.

Examples julia-repl julia> m1 = ExtremeLearner(x, y, 10, σ) Extreme Learning Machine with 10 hidden neurons julia> f1 = fit(m1, sigmoid) [-4.403356409043448, -5.577616954029608, -2.1732800642523595, 0.9669137012255704, -3.6474913410560013, -4.206228346376102, -7.575391282978456, 4.528774205936467, -2.4741301876094655, 40.642730531608635, -11.058942121275233] julia> predictcounterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0]) [9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978] julia> placebotest(m1) ([9.811656638113011e-16, 0.9999999999999962, -9.020553785284482e-17, 0.9999999999999978], [0.5, 0.4, 0.3, 0.2])