ATE/ATT/ITT Estimation

Structs and methods to estimate the ATE, ATT, ITT, and change over time from observational research designs.

CausalELM.Estimators — Module

Estimate causal effects with interrupted time series analysis, G-computation, and doubly robust estimation using Extreme Learning machines.

CausalELM.Estimators.CausalEstimator — Type

Abstract type for GComputation and DoublyRobust

CausalELM.Estimators.InterruptedTimeSeries — Type

Container for the results of an interrupted time series analysis

CausalELM.Estimators.GComputation — Type

Container for the results of G-Computation

CausalELM.Estimators.DoublyRobust — Type

Container for the results of doubly robust estimation

CausalELM.Metalearners — Module

Metalearners to estimate the conditional average treatment effect (CATE).

CausalELM.Metalearners.Metalearner — Type

Abstract type for metalearners

CausalELM.Metalearners.SLearner — Type

S-Learner for CATE estimation.

CausalELM.Metalearners.TLearner — Type

T-Learner for CATE estimation.

CausalELM.Metalearners.XLearner — Type

X-Learner for CATE estimation.

Missing docstring.

Missing docstring for CausalELM.Estimators.estimatecausaleffect!. Check Documenter's build log for details.