CausalELM

Most of the methods and structs here are private, not exported, should not be called by the user, and are documented for the purpose of developing CausalELM or to facilitate understanding of the implementation.

Types

CausalELM.InterruptedTimeSeries — Type

InterruptedTimeSeries(X₀, Y₀, X₁, Y₁; kwargs...)

Initialize an interrupted time series estimator.

Arguments

X₀::Any: an array or DataFrame of covariates from the pre-treatment period.

Y₁::Any: an array or DataFrame of outcomes from the pre-treatment period.

X₁::Any: an array or DataFrame of covariates from the post-treatment period.

Y₁::Any: an array or DataFrame of outcomes from the post-treatment period.

regularized::Function=true: whether to use L2 regularization

Keywords

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For a simple linear regression-based tutorial on interrupted time series analysis see: Bernal, James Lopez, Steven Cummins, and Antonio Gasparrini. "Interrupted time series regression for the evaluation of public health interventions: a tutorial." International journal of epidemiology 46, no. 1 (2017): 348-355.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
m1 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁)
m2 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁; regularized=false)
x₀_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100))
y₀_df = DataFrame(y=rand(100))
x₁_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100)) 
y₁_df = DataFrame(y=rand(100))
m3 = InterruptedTimeSeries(x₀_df, y₀_df, x₁_df, y₁_df)

CausalELM.GComputation — Type

GComputation(X, T, Y; kwargs...)

Initialize a G-Computation estimator.

Arguments

X::Any: an array or DataFrame of covariates.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

task::String: either regression or classification.

quantity_of_interest::String: ATE for average treatment effect or CTE for cummulative treatment effect.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For a good overview of G-Computation see: Chatton, Arthur, Florent Le Borgne, Clémence Leyrat, Florence Gillaizeau, Chloé Rousseau, Laetitia Barbin, David Laplaud, Maxime Léger, Bruno Giraudeau, and Yohann Foucher. "G-computation, propensity score-based methods, and targeted maximum likelihood estimator for causal inference with different covariates sets: a comparative simulation study." Scientific reports 10, no. 1 (2020): 9219.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
m1 = GComputation(X, T, Y)
m2 = GComputation(X, T, Y; task="regression")
m3 = GComputation(X, T, Y; task="regression", quantity_of_interest="ATE)
m4 = GComputation(X, T, Y; task="regression", quantity_of_interest="ATE", regularized=true)

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100)) 
m5 = GComputation(x_df, t_df, y_df)

CausalELM.DoubleMachineLearning — Type

DoubleMachineLearning(X, T, Y; kwargs...)

Initialize a double machine learning estimator with cross fitting.

Arguments

X::Any: an array or DataFrame of covariates of interest.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

W::Any: an array or dataframe of all possible confounders.

task::String: either regression or classification.

quantity_of_interest::String: ATE for average treatment effect or CTE for cummulative treatment effect.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

Unlike other estimators, this method does not support time series or panel data. This method also does not work as well with smaller datasets because it estimates separate outcome models for the treatment and control groups.

References

For more information see: Chernozhukov, Victor, Denis Chetverikov, Mert Demirer, Esther Duflo, Christian Hansen, Whitney Newey, and James Robins. "Double/debiased machine learning for treatment and structural parameters." (2016): C1-C68.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = DoubleMachineLearning(X, T, Y)
m2 = DoubleMachineLearning(X, T, Y; task="regression")

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100))
m3 = DoubleMachineLearning(x_df, t_df, y_df)

CausalELM.SLearner — Type

SLearner(X, T, Y; kwargs...)

Initialize a S-Learner.

Arguments

X::Any: an array or DataFrame of covariates.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

task::String: either regression or classification.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For an overview of S-Learners and other metalearners see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = SLearner(X, T, Y)
m2 = SLearner(X, T, Y; task="regression")
m3 = SLearner(X, T, Y; task="regression", regularized=true)

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100))
m4 = SLearner(x_df, t_df, y_df)

CausalELM.TLearner — Type

TLearner(X, T, Y; kwargs...)

Initialize a T-Learner.

Arguments

X::Any: an array or DataFrame of covariates.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

task::String: either regression or classification.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For an overview of T-Learners and other metalearners see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = TLearner(X, T, Y)
m2 = TLearner(X, T, Y; task="regression")
m3 = TLearner(X, T, Y; task="regression", regularized=true)

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100))
m4 = TLearner(x_df, t_df, y_df)

CausalELM.RLearner — Type

RLearner(X, T, Y; kwargs...)

Initialize an R-Learner.

Arguments

X::Any: an array or DataFrame of covariates of interest.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

W::Any : an array of all possible confounders.

task::String: either regression or classification.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For an explanation of R-Learner estimation see: Nie, Xinkun, and Stefan Wager. "Quasi-oracle estimation of heterogeneous treatment effects." Biometrika 108, no. 2 (2021): 299-319.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = RLearner(X, T, Y)
m2 = RLearner(X, T, Y; t_cat=true)

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100))
m4 = RLearner(x_df, t_df, y_df)

w = rand(100, 6)
m5 = RLearner(X, T, Y, W=w)

CausalELM.CausalEstimator — Type

Abstract type for GComputation and DoubleMachineLearning

CausalELM.Metalearner — Type

Abstract type for metalearners

CausalELM.ExtremeLearningMachine — Type

Abstract type that includes vanilla and L2 regularized Extreme Learning Machines

CausalELM.ExtremeLearner — Type

ExtremeLearner(X, Y, hidden_neurons, activation)

Construct an ExtremeLearner for fitting and prediction.

Notes

While it is possible to use an ExtremeLearner for regression, it is recommended to use RegularizedExtremeLearner, which imposes an L2 penalty, to reduce multicollinearity.

References

For more details see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

See also 'RegularizedExtremeLearner'.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = ExtremeLearner(x, y, 10, σ)

CausalELM.RegularizedExtremeLearner — Type

RegularizedExtremeLearner(X, Y, hidden_neurons, activation)

Construct a RegularizedExtremeLearner for fitting and prediction.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = RegularizedExtremeLearner(x, y, 10, σ)

CausalELM.Nonbinary — Type

Abstract type used to dispatch risk_ratio on nonbinary treatments

CausalELM.Binary — Type

Type used to dispatch risk_ratio on binary treatments

CausalELM.Count — Type

Type used to dispatch risk_ratio on count treatments

CausalELM.Continuous — Type

Type used to dispatch risk_ratio on continuous treatments

Activation Functions

CausalELM.binary_step — Function

binary_step(x)

Apply the binary step activation function to a real number.

Examples

binary_step(1)

binary_step(x)

Apply the binary step activation function to an array.

Examples

binary_step([-1000, 100, 1, 0, -0.001, -3])

CausalELM.σ — Function

σ(x)

Apply the sigmoid activation function to a real number.

Examples

σ(1)

σ(x)

Apply the sigmoid activation function to an array.

Examples

σ([1, 0])

Base.tanh — Function

tanh(x)

Apply the tanh activation function to an array.

Notes

This is just a vectorized version of Base.tanh

Examples

tanh([1, 0])

CausalELM.relu — Function

relu(x)

Apply the ReLU activation function to a real number.

Examples

relu(1)

relu(x)

Apply the ReLU activation function to an array.

Examples

relu([1, 0, -1])

CausalELM.leaky_relu — Function

leaky_relu(x)

Apply the leaky ReLU activation function to a real number.

Examples

leaky_relu(1)

leaky_relu(x)

Apply the leaky ReLU activation function to an array.

Examples

leaky_relu([-0.01, 0, 1])

CausalELM.swish — Function

swish(x)

Apply the swish activation function to a real number.

Examples

swish(1)

swish(x)

Apply the swish activation function to an array.

Examples

swish([1, 0, -1])

CausalELM.softmax — Function

softmax(x)

Apply the softmax activation function to a real number.

Examples

softmax(1)

softmax(x)

Apply the softmax activation function to a vector.

Examples

softmax([1, 2, 3])

softmax(x)

Apply the softmax activation function to the rows of an array.

Examples

julia> x = rand(5, 3)
softmax(x)

CausalELM.softplus — Function

softplus(x)

Apply the softplus activation function to a real number.

Examples

softplus(1)

softplus(x)

Apply the softplus activation function to an array.

Examples

softplus([1, -1])

CausalELM.gelu — Function

gelu(x)

Apply the GeLU activation function to a real number.

Examples

gelu(1)

gelu(x)

Apply the GeLU activation function to an array.

Examples

gelu([-1, 0, 1])

CausalELM.gaussian — Function

gaussian(x)

Apply the gaussian activation function to a real number.

Examples

gaussian(1)

gaussian(x)

Apply the gaussian activation function to an array.

Examples

gaussian([1, -1])

CausalELM.hard_tanh — Function

hard_tanh(x)

Apply the hard_tanh activation function to a real number.

Examples

hard_tanh(-2)

hard_tanh(x)

Apply the hard_tanh activation function to an array.

Examples

hard_tanh([-2, 0, 2])

CausalELM.elish — Function

elish(x)

Apply the ELiSH activation function to a real number.

Examples

elish(1)

elish(x)

Apply the ELiSH activation function to an array.

Examples

elish([-1, 1])

CausalELM.fourier — Function

fourrier(x)

Apply the Fourier activation function to a real number.

Examples

fourier(1)

fourrier(x)

Apply the Fourier activation function to an array.

Examples

fourier([-1, 1])

Cross Validation

CausalELM.generate_folds — Function

generate_folds(X, Y, folds)

Create folds for cross validation.

Examples

xfolds, y_folds = generate_folds(zeros(20, 2), zeros(20), 5)

CausalELM.generate_temporal_folds — Function

generate_folds(X, Y, folds)

Create rolling folds for cross validation of time series data.

Examples

xfolds, y_folds = generate_temporal_folds(zeros(20, 2), zeros(20), 5, temporal=true)

CausalELM.validation_loss — Function

validation_loss(xtrain, ytrain, xtest, ytest, nodes, metric; kwargs...)

Calculate a validation metric for a single fold in k-fold cross validation.

Arguments

xtrain::Any: an array of features to train on.

ytrain::Any: an array of training labels.

xtest::Any: an array of features to test on.

ytrain::Any: an array of testing labels.

nodes::Int: the number of neurons in the extreme learning machine.

metric::Function: the validation metric to calculate.

Keywords

activation::Function=relu: the activation function to use.

regularized::Function=true: whether to use L2 regularization.

Examples

x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
validation_loss(x, y, 5, accuracy, 3)

CausalELM.cross_validate — Function

cross_validate(X, Y, neurons, metric, activation, regularized, folds, temporal)

Calculate a validation metric for k folds using a single set of hyperparameters.

Arguments

X::Array: an array of features to train on.

Y::Vector: a vector of labels to train on.

neurons::Int: the number of neurons to use in the extreme learning machine.

metric::Function: the validation metric to calculate.

activation::Function=relu: the activation function to use.

regularized::Function=true: whether to use L2 regularization

folds::Int: the number of folds to use for cross validation.

temporal::Function=true: whether the data is of a time series or panel nature.

Examples

x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
cross_validate(x, y, 5, accuracy)

CausalELM.best_size — Function

best_size(X, Y, metric, task, activation, min_neurons, max_neurons, regularized, folds, 
          temporal, iterations, elm_size)

Compute the best number of neurons for an Extreme Learning Machine.

Notes

The procedure tests networks with numbers of neurons in a sequence whose length is given by iterations on the interval [minneurons, maxneurons]. Then, it uses the networks sizes and validation errors from the sequence to predict the validation error or metric for every network size between minneurons and maxneurons using the function approximation ability of an Extreme Learning Machine. Finally, it returns the network size with the best predicted validation error or metric.

Arguments

X::Array: an array of features to train on.

Y::Vector: a vector of labels to train on.

metric::Function: the validation metric to calculate.

task::String: either regression or classification.

activation::Function=relu: the activation function to use.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

regularized::Function=true: whether to use L2 regularization

folds::Int: the number of folds to use for cross validation.

temporal::Function=true: whether the data is of a time series or panel nature.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

elm_size::Int: the number of nuerons in the validation loss approximator network.

Examples

best_size(rand(100, 5), rand(100), mse, "regression")

CausalELM.shuffle_data — Function

shuffle_data(X, Y)

Shuffles covariates and outcome vector for cross validation.

Examples

x, y, t = rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
shuffle_data(x, y, t)

Average Causal Effect Estimators

CausalELM.XLearner — Type

XLearner(X, T, Y; kwargs...)

Initialize an X-Learner.

Arguments

X::Any: an array or DataFrame of covariates.

T::Any: an vector or DataFrame of treatment statuses.

Y::Any: an array or DataFrame of outcomes.

Keywords

task::String: either regression or classification.

regularized::Function=true: whether to use L2 regularization

activation::Function=relu: the activation function to use.

validation_metric::Function: the validation metric to calculate during cross validation.

min_neurons::Int: the minimum number of neurons to consider for the extreme learner.

max_neurons::Int: the maximum number of neurons to consider for the extreme learner.

folds::Int: the number of cross validation folds to find the best number of neurons.

iterations::Int: the number of iterations to perform cross validation between minneurons and maxneurons.

approximator_neurons::Int: the number of nuerons in the validation loss approximator network.

Notes

If regularized is set to true then the ridge penalty will be estimated using generalized cross validation where the maximum number of iterations is 2 * folds for the successive halving procedure. However, if the penalty in on iteration is approximately the same as in the previous penalty, then the procedure will stop early.

References

For an overview of X-Learners and other metalearners see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

For details and a derivation of the generalized cross validation estimator see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = XLearner(X, T, Y)
m2 = XLearner(X, T, Y; task="regression")
m3 = XLearner(X, T, Y; task="regression", regularized=true)

x_df = DataFrame(x1=rand(100), x2=rand(100), x3=rand(100), x4=rand(100))
t_df, y_df = DataFrame(t=rand(0:1, 100)), DataFrame(y=rand(100))
m4 = XLearner(x_df, t_df, y_df)

CausalELM.estimate_effect! — Function

estimate_effect!(DML, [,cate])

Estimate a treatment effect using double machine learning.

Notes

This method should not be called directly.

Arguments

DML::DoubleMachineLearning: the DoubleMachineLearning struct to estimate the effect for.

cate::Bool=false: whether to estimate the cate.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = DoubleMachineLearning(X, T, Y)
estimate_effect!(m1)

estimate_effect!(DRE, X, T, Y, Z)

Estimate the CATE for a single cross fitting iteration via doubly robust estimation.

This method should not be called directly.

Arguments

DRE::DoublyRobustLearner: the DoubleMachineLearning struct to estimate the effect for.

X: a vector of three covariate folds.

T: a vector of three treatment folds.

Y: a vector of three outcome folds.

Z : a vector of three confounder folds and covariate folds.

Examples

X, T, Y, W =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100), rand(6, 100)
m1 = DoublyRobustLearner(X, T, Y, W=W)

X, T, W, Y = make_folds(m1)
Z = m1.W == m1.X ? X : [reduce(hcat, (z)) for z in zip(X, W)]
estimate_effect!(m1, X, T, Y, Z)

CausalELM.predict_residuals — Function

predict_residuals(DML, x_train, x_test, y_train, y_test, t_train, t_test)

Predict treatment and outcome residuals for doubl machine learning.

Notes

This method should not be called directly.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
x_train, x_test = X[1:80, :], X[81:end, :]
y_train, y_test = Y[1:80], Y[81:end]
t_train, t_test = T[1:80], T[81:100]
m1 = DoubleMachineLearning(X, T, Y)
predict_residuals(m1, x_train, x_test, y_train, y_test, t_train, t_test)

CausalELM.moving_average — Function

moving_average(x)

Calculates a cumulative moving average.

Examples

moving_average([1, 2, 3])

Metalearners

CausalELM.stage1! — Function

stage1!(x)

Estimate the first stage models for an X-learner.

This method should not be called by the user.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = XLearner(X, T, Y)
stage1!(m1)

CausalELM.stage2! — Function

stage2!(x)

Estimate the second stage models for an X-learner.

This method should not be called by the user.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = XLearner(X, T, Y)
stage1!(m1)
stage2!(m1)

Common Methods

CausalELM.estimate_causal_effect! — Function

estimate_causal_effect!(its)

Estimate the effect of an event relative to a predicted counterfactual.

Examples

X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
m1 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁)
estimate_causal_effect!(m1)

estimate_causal_effect!(g)

Estimate a causal effect of interest using G-Computation.

Notes

If treatents are administered at multiple time periods, the effect will be estimated as the average difference between the outcome of being treated in all periods and being treated in no periods. For example, given that ividuals 1, 2, ..., i ∈ I recieved either a treatment or a placebo in p different periods, the model would estimate the average treatment effect as E[Yᵢ|T₁=1, T₂=1, ... Tₚ=1, Xₚ] - E[Yᵢ|T₁=0, T₂=0, ... Tₚ=0, Xₚ].

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = GComputation(X, T, Y)
estimate_causal_effect!(m1)

estimate_causal_effect!(DML)

Estimate a causal effect of interest using double machine learning.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = DoubleMachineLearning(X, T, Y)
estimate_causal_effect!(m1)

W = rand(100, 6)
m2 = DoubleMachineLearning(X, T, Y, W=W)
estimate_causal_effect!(m2)

estimate_causal_effect!(s)

Estimate the CATE using an S-learner.

For an overview of S-learning see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m4 = SLearner(X, T, Y)
estimate_causal_effect!(m4)

estimate_causal_effect!(t)

Estimate the CATE using an T-learner.

For an overview of T-learning see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m5 = TLearner(X, T, Y)
estimate_causal_effect!(m5)

estimate_causal_effect!(x)

Estimate the CATE using an X-learner.

For an overview of X-learning see: Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for estimating heterogeneous treatment effects using machine learning." Proceedings of the national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = XLearner(X, T, Y)
estimate_causal_effect!(m1)

estimate_causal_effect!(R)

Estimate the CATE using an R-learner.

For an overview of R-learning see: Nie, Xinkun, and Stefan Wager. "Quasi-oracle estimation of heterogeneous treatment effects." Biometrika 108, no. 2 (2021): 299-319.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = RLearner(X, T, Y)
estimate_causal_effect!(m1)

estimate_causal_effect!(DRE)

Estimate the CATE using a doubly robust learner.

For details on how this method estimates the CATE see: Kennedy, Edward H. "Towards optimal doubly robust estimation of heterogeneous causal effects." Electronic Journal of Statistics 17, no. 2 (2023): 3008-3049.

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = DoublyRobustLearner(X, T, Y)
estimate_causal_effect!(m1)

Inference

CausalELM.summarize — Function

summarize(mod, n)

Get a summary from a CausalEstimator or Metalearner.

Arguments

mod::Union{CausalEstimator, Metalearner}: a model to summarize.

n::Int=100: the number of iterations to generate the numll distribution for randomization inference.

Notes

p-values and standard errors are estimated using approximate randomization inference.

References

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

X, T, Y =  rand(100, 5), [rand()<0.4 for i in 1:100], rand(100)
m1 = GComputation(X, T, Y)
estimate_causal_effect!(m1)
summarize(m1)

m2 = RLearner(X, T, Y)
estimate_causal_effect(m2)
julia> summarize(m2)

m3 = SLearner(X, T, Y)
estimate_causal_effect!(m3)
summarise(m3)

summarize(its, n, mean_effect)

Get a summary from an interrupted time series estimator.

Arguments

its::InterruptedTimeSeries: an interrupted time series estimator

n::Int=100: the number of iterations to generate the numll distribution for randomization inference.

mean_effect::Bool=true: whether to estimate the mean or cumulative effect for an interrupted time series estimator.

Examples

X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
m4 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁)
estimate_causal_effect!(m4)
julia> summarize(m4)

CausalELM.generate_null_distribution — Function

generate_null_distribution(mod, n)

Generate a null distribution for the treatment effect of G-computation, double machine learning, or metalearning.

Arguments

mod::Union{CausalEstimator, Metalearner}: a model to summarize.

n::Int=100: the number of iterations to generate the numll distribution for randomization inference.

Notes

This method estimates the same model that is provided using random permutations of the treatment assignment to generate a vector of estimated effects under different treatment regimes. When mod is a metalearner the null statistic is the difference is the ATE.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

Examples

x, t, y = rand(100, 5), [rand()<0.4 for i in 1:100], rand(1:100, 100, 1)
g_computer = GComputation(x, t, y)
estimate_causal_effect!(g_computer)
generate_null_distribution(g_computer, 500)

generate_null_distribution(its, n, mean_effect)

Arguments

its::InterruptedTimeSeries: an interrupted time series estimator

n::Int=100: the number of iterations to generate the numll distribution for randomization inference.

mean_effect::Bool=true: whether to estimate the mean or cumulative effect for an interrupted time series estimator.

Examples

x₀, y₀, x₁, y₁ = rand(1:100, 100, 5), rand(100), rand(10, 5), rand(10)
its = InterruptedTimeSeries(x₀, y₀, x₁, y₁)
estimate_causale_ffect!(its)
generate_null_distribution(its, 10)

CausalELM.quantities_of_interest — Function

quantities_of_interest(mod, n)

Generate a p-value and standard error through randomization inference

This method generates a null distribution of treatment effects by reestimating treatment effects from permutations of the treatment vector and estimates a p-value and standard from the generated distribution.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

x, t, y = rand(100, 5), [rand()<0.4 for i in 1:100], rand(1:100, 100, 1)
g_computer = GComputation(x, t, y)
estimate_causal_effect!(g_computer)
quantities_of_interest(g_computer, 1000)

quantities_of_interest(mod, n)

Generate a p-value and standard error through randomization inference

This method generates a null distribution of treatment effects by reestimating treatment effects from permutations of the treatment vector and estimates a p-value and standard from the generated distribution. Randomization for event studies is done by creating time splits at even intervals and reestimating the causal effect.

Note that lowering the number of iterations increases the probability of failing to reject the null hypothesis.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

x₀, y₀, x₁, y₁ = rand(1:100, 100, 5), rand(100), rand(10, 5), rand(10)
its = InterruptedTimeSeries(x₀, y₀, x₁, y₁)
estimate_causal_effect!(its)
quantities_of_interest(its, 10)

Model Validation

CausalELM.validate — Function

validate(its; kwargs...)

Test the validity of an estimated interrupted time series analysis.

Arguments

its::InterruptedTimeSeries: an interrupted time seiries estimator.

Keywords

n::Int: the number of times to simulate a confounder.

low::Float64=0.15: the minimum proportion of data points to include before or after the tested break in the Wald supremum test.

high::Float64=0.85: the maximum proportion of data points to include before or after the tested break in the Wald supremum test.

Notes

This method coducts a Chow Test, a Wald supremeum test, and tests the model's sensitivity to confounders. The Chow Test tests for structural breaks in the covariates between the time before and after the event. p-values represent the proportion of times the magnitude of the break in a covariate would have been greater due to chance. Lower p-values suggest a higher probability the event effected the covariates and they cannot provide unbiased counterfactual predictions. The Wald supremum test finds the structural break with the highest Wald statistic. If this is not the same as the hypothesized break, it could indicate an anticipation effect, a confounding event, or that the intervention or policy took place in multiple phases. p-values represent the proportion of times we would see a larger Wald statistic if the data points were randomly allocated to pre and post-event periods for the predicted structural break. Ideally, the hypothesized break will be the same as the predicted break and it will also have a low p-value. The omitted predictors test adds normal random variables with uniform noise as predictors. If the included covariates are good predictors of the counterfactual outcome, adding irrelevant predictors should not have a large effect on the predicted counterfactual outcomes or the estimated effect.

This method does not implement the second test in Baicker and Svoronos because the estimator in this package models the relationship between covariates and the outcome and uses an extreme learning machine instead of linear regression, so variance in the outcome across different bins is not much of an issue.

References

For more details on the assumptions and validity of interrupted time series designs, see: Baicker, Katherine, and Theodore Svoronos. Testing the validity of the single interrupted time series design. No. w26080. National Bureau of Economic Research, 2019.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
m1 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁)
stimate_causal_effect!(m1)
julia> validate(m1)

validate(m; kwargs)

Arguments

m::Union{CausalEstimator, Metalearner}: a model to validate/test the assumptions of.

Keywords

num_treatments=5::Int: the maximum number of treatments to use when testing the plausability of the counterfactual consistency assumption.

min::Float64=1.0e-6: minimum probability of treatment for the positivity assumption.

high::Float64=1-min: the maximum probability of treatment for the positivity assumption.

Notes

This method tests the counterfactual consistency, exchangeability, and positivity assumptions required for causal inference. It should be noted that consistency and exchangeability are not directly testable, so instead, these tests do not provide definitive evidence of a violation of these assumptions. To probe the counterfactual consistency assumption, we assume there were multiple levels of treatments and find them by binning the dependent vairable for treated observations using Jenks breaks. The optimal number of breaks between 2 and num_treatments is found using the elbow method. Using these hypothesized treatment assignemnts, this method compares the MSE of linear regressions using the observed and hypothesized treatments. If the counterfactual consistency assumption holds then the difference between the MSE with hypothesized treatments and the observed treatments should be positive because the hypothesized treatments should not provide useful information. If it is negative, that indicates there was more useful information provided by the hypothesized treatments than the observed treatments or that there is an unobserved confounder. Next, this methods tests the model's sensitivity to a violation of the exchangeability assumption by calculating the E-value, which is the minimum strength of association, on the risk ratio scale, that an unobserved confounder would need to have with the treatment and outcome variable to fully explain away the estimated effect. Thus, higher E-values imply the model is more robust to a violation of the exchangeability assumption. Finally, this method tests the positivity assumption by estimating propensity scores. Rows in the matrix are levels of covariates that have a zero probability of treatment. If the matrix is empty, none of the observations have an estimated zero probability of treatment, which implies the positivity assumption is satisfied.

References

For a thorough review of casual inference assumptions see: Hernan, Miguel A., and James M. Robins. Causal inference what if. Boca Raton: Taylor and Francis, 2024.

For more information on the E-value test see: VanderWeele, Tyler J., and Peng Ding. "Sensitivity analysis in observational research: introducing the E-value." Annals of internal medicine 167, no. 4 (2017): 268-274.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100]), vec(rand(1:100, 100, 1)) 
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
validate(g_computer)

CausalELM.covariate_independence — Function

covariate_independence(its; kwargs..)

Test for independence between covariates and the event or intervention.

Arguments

its::InterruptedTImeSeries: an interrupted time seiries estimator.

Keywords

n::Int: the number of permutations for assigning observations to the pre and post-treatment periods.

This is a Chow Test for covariates with p-values estimated via randomization inference. The p-values are the proportion of times randomly assigning observations to the pre or post-intervention period would have a larger estimated effect on the the slope of the covariates. The lower the p-values, the more likely it is that the event/intervention effected the covariates and they cannot provide an unbiased prediction of the counterfactual outcomes.

For more information on using a Chow Test to test for structural breaks see: Baicker, Katherine, and Theodore Svoronos. Testing the validity of the single interrupted time series design. No. w26080. National Bureau of Economic Research, 2019.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

x₀, y₀, x₁, y₁ = (Float64.(rand(1:5, 100, 5)), randn(100), rand(1:5, (10, 5)), randn(10))
its = InterruptedTimeSeries(x₀, y₀, x₁, y₁)
estimate_causal_effect!(its)
covariate_independence(its)

CausalELM.omitted_predictor — Function

omitted_predictor(its; kwargs...)

See how an omitted predictor/variable could change the results of an interrupted time series analysis.

Arguments

its::InterruptedTImeSeries: an interrupted time seiries estimator.

Keywords

n::Int: the number of times to simulate a confounder.

Notes

This method reestimates interrupted time series models with uniform random variables. If the included covariates are good predictors of the counterfactual outcome, adding a random variable as a covariate should not have a large effect on the predicted counterfactual outcomes and therefore the estimated average effect.

References

For more information on using a Chow Test to test for structural breaks see: Baicker, Katherine, and Theodore Svoronos. Testing the validity of the single interrupted time series design. No. w26080. National Bureau of Economic Research, 2019.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

x₀, y₀, x₁, y₁ = (Float64.(rand(1:5, 100, 5)), randn(100), rand(1:5, (10, 5)), randn(10))
its = InterruptedTimeSeries(x₀, y₀, x₁, y₁)
estimate_causal_effect!(its)
omitted_predictor(its)

CausalELM.sup_wald — Function

sup_wald(its; kwargs)

Check if the predicted structural break is the hypothesized structural break.

Arguments

its::InterruptedTimeSeries: an interrupted time seiries estimator.

Keywords

n::Int: the number of times to simulate a confounder.

low::Float64=0.15: the minimum proportion of data points to include before or after the tested break in the Wald supremum test.

high::Float64=0.85: the maximum proportion of data points to include before or after the tested break in the Wald supremum test.

Notes

This method conducts Wald tests and identifies the structural break with the highest Wald statistic. If this break is not the same as the hypothesized break, it could indicate an anticipation effect, confounding by some other event or intervention, or that the intervention or policy took place in multiple phases. p-values are estimated using approximate randomization inference and represent the proportion of times we would see a larger Wald statistic if the data points were randomly allocated to pre and post-event periods for the predicted structural break.

References

For more information on using a Chow Test to test for structural breaks see: Baicker, Katherine, and Theodore Svoronos. Testing the validity of the single interrupted time series design. No. w26080. National Bureau of Economic Research, 2019.

For a primer on randomization inference see: https://www.mattblackwell.org/files/teaching/s05-fisher.pdf

Examples

x₀, y₀, x₁, y₁ = (Float64.(rand(1:5, 100, 5)), randn(100), rand(1:5, (10, 5)), randn(10))
its = InterruptedTimeSeries(x₀, y₀, x₁, y₁)
estimate_causal_effect!(its)
sup_wald(its)

CausalELM.p_val — Function

p_val(x, y, β; kwargs...)

Estimate the p-value for the hypothesis that an event had a statistically significant effect on the slope of a covariate using randomization inference.

Arguments

x::Array{<:Real}: covariates.

y::Array{<:Real}: the outcome.

β::Array{<:Real}=0.15: the fitted weights.

Keywords

two_sided::Bool=false: whether to conduct a one-sided hypothesis test.

Examples

x, y, β = reduce(hcat, (float(rand(0:1, 10)), ones(10))), rand(10), 0.5
p_val(x, y, β)
p_val(x, y, β; n=100, two_sided=true)

CausalELM.counterfactual_consistency — Function

counterfactual_consistency(m; kwargs...)

Arguments

m::Union{CausalEstimator, Metalearner}: a model to validate/test the assumptions of.

Keywords

num_treatments=5::Int: the maximum number of treatments to use when testing the plausability of the counterfactual consistency assumption.

Notes

Examine the counterfactual consistency assumption. First, this function generates Jenks breaks based on outcome values for the treatment group. Then, it replaces treatment statuses with the numbers corresponding to each group. Next, it runs two models, one for the treatment group, one with and one without the fake treatment assignemnts generated by the Jenks breaks. Finally, it subtracts the mean squared error from the regression with real data from the mean squared error from the regression with the fake treatment statuses. If this number is negative, it might indicate a violation of the counterfactual consistency assumption or omitted variable bias.

References

For a primer on G-computation and its assumptions see: Naimi, Ashley I., Stephen R. Cole, and Edward H. Kennedy. "An introduction to g methods." International journal of epidemiology 46, no. 2 (2017): 756-762.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100], vec(rand(1:100, 100, 1)))
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
counterfactual_consistency(g_computer)

CausalELM.exchangeability — Function

exchangeability(model)

Test the sensitivity of a G-computation or doubly robust estimator or metalearner to a violation of the exchangeability assumption.

References

For more information on the E-value test see: VanderWeele, Tyler J., and Peng Ding. "Sensitivity analysis in observational research: introducing the E-value." Annals of internal medicine 167, no. 4 (2017): 268-274.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100], vec(rand(1:100, 100, 1)))
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
e_value(g_computer)

CausalELM.e_value — Function

e_value(model)

Test the sensitivity of an estimator to a violation of the exchangeability assumption.

References

For more information on the E-value test see: VanderWeele, Tyler J., and Peng Ding. "Sensitivity analysis in observational research: introducing the E-value." Annals of internal medicine 167, no. 4 (2017): 268-274.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100], vec(rand(1:100, 100, 1)))
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
e_value(g_computer)

CausalELM.binarize — Function

binarize(x, cutoff)

Convert a vector of counts or a continuous vector to a binary vector.

Examples

binarize([1, 2, 3], 2)

CausalELM.risk_ratio — Function

risk_ratio(model)

Calculate the risk ratio for an estimated model.

Notes

If the treatment variable is not binary and the outcome variable is not continuous then the treatment variable will be binarized.

References

For more information on how other quantities of interest are converted to risk ratios see: VanderWeele, Tyler J., and Peng Ding. "Sensitivity analysis in observational research: introducing the E-value." Annals of internal medicine 167, no. 4 (2017): 268-274.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100], vec(rand(1:100, 100, 1)))
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
risk_ratio(g_computer)

CausalELM.positivity — Function

positivity(model, [,min], [,max])

Find likely violations of the positivity assumption.

Notes

This method uses an extreme learning machine or regularized extreme learning machine to estimate probabilities of treatment. The returned matrix, which may be empty, are the covariates that have a (near) zero probability of treatment or near zero probability of being assigned to the control group, whith their entry in the last column being their estimated treatment probability. In other words, they likely violate the positivity assumption.

Arguments

model::Union{CausalEstimator, Metalearner}: a model to validate/test the assumptions of.

min::Float64=1.0e-6: minimum probability of treatment for the positivity assumption.

high::Float64=1-min: the maximum probability of treatment for the positivity assumption.

Examples

x, t, y = rand(100, 5), Float64.([rand()<0.4 for i in 1:100], vec(rand(1:100, 100, 1)))
g_computer = GComputation(x, t, y, temporal=false)
estimate_causal_effect!(g_computer)
positivity(g_computer)

CausalELM.sums_of_squares — Function

sums_of_squares(data, num_classes)

Calculate the minimum sum of squares for each data point and class for the Jenks breaks algorithm.

Notes

This should not be called by the user.

Examples

sums_of_squares([1, 2, 3, 4, 5], 2)

CausalELM.class_pointers — Function

class_pointers(data, num_classes, sums_of_sqs)

Compute class pointers that minimize the sum of squares for Jenks breaks.

Notes

This should not be callled by the user.

Examples

sums_squares = sums_of_squares([1, 2, 3, 4, 5], 2)
class_pointers([1, 2, 3, 4, 5], 2, sums_squares)

CausalELM.backtrack_to_find_breaks — Function

backtrack_to_find_breaks(data, num_classes, sums_of_sqs)

Determine break points from class assignments.

Notes

This should not be called by the user.

Examples

data = [1, 2, 3, 4, 5]
ptr = class_pointers([1, 2, 3, 4, 5], 2, sums_of_squares([1, 2, 3, 4, 5], 2))
backtrack_to_find_breaks([1, 2, 3, 4, 5], ptr)

CausalELM.variance — Function

variance(data)

Calculate the variance of some numbers.

Notes

This function does not use Besel's correction.

Examples

variance([1, 2, 3, 4, 5])

CausalELM.best_splits — Function

best_splits(data, num_classes)

Find the best number of splits for Jenks breaks.

Notes

This function finds the best number of splits by finding the number of splits that results in the greatest decrease in the slope of the line between itself and its GVF and the next higher number of splits and its GVF. This is the same thing as the elbow method.

This should nto be called by the user.

Examples

best_splits(collect(1:10), 5)

CausalELM.group_by_class — Function

group_by_class(data, classes)

Group data points into vectors such that data points assigned to the same class are in the same vector.

Notes

This should nto be called by the user.

Examples

group_by_class([1, 2, 3, 4, 5], [1, 1, 1, 2, 3])

CausalELM.jenks_breaks — Function

jenks_breaks(data, num_classes)

Generate Jenks breaks for a vector of real numbers.

Examples

jenks_breaks([1, 2, 3, 4, 5], 3)

CausalELM.fake_treatments — Function

fake_treatments(data, num_classes)

Generate fake treatment statuses corresponding to the classes assigned by the Jenks breaks algorithm.

Examples

fake_treatments([1, 2, 3, 4, 5], 4)

CausalELM.sdam — Function

sdam(x)

Calculate the sum of squared deviations for array mean for a set of sub arrays.

Examples

sdam([5, 4, 9, 10])

CausalELM.scdm — Function

sdcm(x)

Calculate the sum of squared deviations for class means for a set of sub arrays.

Examples

CausalELM.gvf — Function

gvf(x)

Calculate the goodness of variance fit for a set of sub vectors.

Examples

gvf([[4, 5], [9, 10]])

CausalELM.var_type — Function

var_type(x)

Determine the type of variable held by a vector.

Examples

var_type([1, 2, 3, 2, 3, 1, 1, 3, 2])

Validation Metrics

CausalELM.mse — Function

mse(y, ŷ)

Calculate the mean squared error

See also mae.

Examples

mse([0.0, 0.0, 0.0], [0.0, 0.0, 0.0])
mse([-1.0, -1.0, -1.0], [1.0, 1.0, 1.0])

CausalELM.mae — Function

mae(y, ŷ)

Calculate the mean absolute error

See also mse.

Examples

mae([-1.0, -1.0, -1.0], [1.0, 1.0, 1.0])
mae([1.0, 1.0, 1.0], [2.0, 2.0, 2.0])

CausalELM.accuracy — Function

accuracy(y, ŷ)

Calculate the accuracy for a classification task

Examples

accuracy([1, 1, 1, 1], [0, 1, 1, 0])
accuracy([1, 2, 3, 4], [1, 1, 1, 1])

Base.precision — Function

precision(y, ŷ)

Calculate the precision for a classification task

See also recall.

Examples

precision([0, 1, 0, 0], [0, 1, 1, 0])
precision([0, 1, 0, 0], [0, 1, 0, 0])

CausalELM.recall — Function

recall(y, ŷ)

Calculate the recall for a classification task

See also precision.

Examples

recall([1, 2, 1, 3, 0], [2, 2, 2, 3, 1])
recall([1, 2, 1, 3, 2], [2, 2, 2, 3, 1])

CausalELM.F1 — Function

F1(y, ŷ)

Calculate the F1 score for a classification task

Examples

F1([1, 2, 1, 3, 0], [2, 2, 2, 3, 1])
F1([1, 2, 1, 3, 2], [2, 2, 2, 3, 1])

CausalELM.confusion_matrix — Function

confusion_matrix(y, ŷ)

Generate a confusion matrix

Examples

confusion_matrix([1, 1, 1, 1, 0], [1, 1, 1, 1, 0])
confusion_matrix([1, 1, 1, 1, 0, 2], [1, 1, 1, 1, 0, 2])

Extreme Learning Machines

CausalELM.fit! — Function

fit!(model)

Make predictions with an ExtremeLearner.

References

For more details see: Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = ExtremeLearner(x, y, 10, σ)

fit!(model)

Fit a Regularized Extreme Learner.

References

For more details see: Li, Guoqiang, and Peifeng Niu. "An enhanced extreme learning machine based on ridge regression for regression." Neural Computing and Applications 22, no. 3 (2013): 803-810.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = RegularizedExtremeLearner(x, y, 10, σ)
f1 = fit!(m1)

CausalELM.predict — Function

predict(model, X)

Use an ExtremeLearningMachine to make predictions.

References

For more details see: Huang G-B, Zhu Q-Y, Siew C. Extreme learning machine: theory and applications. Neurocomputing. 2006;70:489–501. https://doi.org/10.1016/j.neucom.2005.12.126

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = ExtremeLearner(x, y, 10, σ)
f1 = fit(m1, sigmoid)
julia> predict(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0])

CausalELM.predict_counterfactual! — Function

predictcounterfactual(model, X)

Use an ExtremeLearningMachine to predict the counterfactual.

Notes

This should be run with the observed covariates. To use synthtic data for what-if scenarios use predict.

See also predict.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = ExtremeLearner(x, y, 10, σ)
f1 = fit(m1, sigmoid)
predict_counterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0])

CausalELM.placebo_test — Function

placebo_test(model)

Conduct a placebo test.

Notes

This method makes predictions for the post-event or post-treatment period using data in the pre-event or pre-treatment period and the post-event or post-treament. If there is a statistically significant difference between these predictions the study design may be flawed. Due to the multitude of significance tests for time series data, this function returns the predictions but does not test for statistical significance.

Examples

x, y = [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0], [0.0, 1.0, 0.0, 1.0]
m1 = ExtremeLearner(x, y, 10, σ)
f1 = fit(m1, sigmoid)
predict_counterfactual(m1, [1.0 1.0; 0.0 1.0; 0.0 0.0; 1.0 0.0])
placebo_test(m1)

CausalELM.ridge_constant — Function

ridge_constant(model, [,iterations])

Calculate the L2 penalty for a regularized extreme learning machine using generalized cross validation with successive halving.

Arguments

model::RegularizedExtremeLearner: a regularized extreme learning machine

iterations::Int: the number of iterations to perform for successive halving.

References

For more information see: Golub, Gene H., Michael Heath, and Grace Wahba. "Generalized cross-validation as a method for choosing a good ridge parameter." Technometrics 21, no. 2 (1979): 215-223.

Examples

m1 = RegularizedExtremeLearner(x, y, 10, σ)
ridge_constant(m1)
ridge_constant(m1, iterations=20)

CausalELM.set_weights_biases — Function

set_weights_biases(model)

Calculate the weights and biases for an extreme learning machine or regularized extreme learning machine.

References

For details see; Huang, Guang-Bin, Qin-Yu Zhu, and Chee-Kheong Siew. "Extreme learning machine: theory and applications." Neurocomputing 70, no. 1-3 (2006): 489-501.

Examples

m1 = RegularizedExtremeLearner(x, y, 10, σ)
set_weights_biases(m1)

Utility Functions

CausalELM.mean — Function

mean(x)

Calculate the mean of a vector.

Examples

mean([1, 2, 3, 4])

CausalELM.var — Function

var(x)

Calculate the (sample) mean of a vector.

Examples

var([1, 2, 3, 4])

CausalELM.consecutive — Function

consecutive(x)

Subtract consecutive elements in a vector.

Notes

This function is only used to create a rolling average for interrupted time series analysis.

Examples

consecutive([1, 2, 3, 4, 5])

CausalELM.one_hot_encode — Function

one_hot_encode(x)

One hot encode a categorical vector for multiclass classification.

Examples

one_hot_encode([1, 2, 3, 4, 5])