Cross Valdiation

Methods to find the optimal number of neurons via cross validation

CausalELM.CrossValidation — Module

Methods to perform cross validation and find the optimum number of neurons.

To reduce computation time, the number of neurons is optimized by using cross validation to estimate the validation error on a small subset of the range of possible numbers of neurons. Then, an Extreme Learning Machine is trained to predict validation loss from the given cross validation sets. Finally, the number of neurons is selected that has the smallest predicted loss or the highest classification metric.

CausalELM.CrossValidation.recode — Function

recode(ŷ)

Round predicted values to their predicted class for classification tasks.

If the smallest predicted label is 0, all labels are shifted up 1; if the smallest label is -1, all labels are shifted up 2. Also labels cannot be smaller than -1.

Examples

julia> recode([-0.7, 0.2, 1.1])
3-element Vector{Float64}
1
2
3
julia> recode([0.1, 0.2, 0.3])
3-element Vector{Float64}
1
1
1
julia> recode([1.1, 1.51, 1.8])
3-element Vector{Float64}
1
2
2

CausalELM.CrossValidation.generate_folds — Function

generate_folds(X, Y, folds)

Creates folds for cross validation.

Examples

julia> xfolds, y_folds = generate_folds(zeros(20, 2), zeros(20), 5)

CausalELM.CrossValidation.validate_fold — Function

validate_fold(xtrain, ytrain, xtest, ytest, nodes, metric; activation, regularized)

Calculate a validation metric for a single fold in k-fold cross validation.

Examples

julia> x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
julia> validate_fold(x, y, 5, accuracy, 3)
0.0

CausalELM.CrossValidation.cross_validate — Function

cross_validate(X, Y, neurons, metric, activation, regularized, folds, temporal)

Calculate a validation metric for k folds using a single set of hyperparameters.

Examples

julia> x = rand(100, 5); y = Float64.(rand(100) .> 0.5)
julia> cross_validate(x, y, 5, accuracy)
0.0257841765251021

CausalELM.CrossValidation.best_size — Function

best_size(X, Y, metric, task, activation, min_neurons, max_neurons, regularized, folds, 
    temporal, iterations, approximator_neurons)

Compute the best number of neurons for an Extreme Learning Machine.

The procedure tests networks with numbers of neurons in a sequence whose length is given by iterations on the interval [minneurons, maxneurons]. Then, it uses the networks sizes and validation errors from the sequence to predict the validation error or metric for every network size between minneurons and maxneurons using the function approximation ability of an Extreme Learning Machine. Finally, it returns the network size with the best predicted validation error or metric.

Examples

julia> best_size(rand(100, 5), rand(100), mse, "regression")
11

CausalELM.CrossValidation.shuffle_data — Function

shuffle_data(X, Y, T)

Shuffles covariates, treatment vector, and outcome vector for cross validation.

Examples

julia> x, y, t = rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> shuffle_data(x, y, t)
([0.6124923085225416 0.2713900065807924 … 0.6094796972512194 0.6067966603192685; 
0.7186612932571539 0.8047878363606299 … 0.9787878554455594 0.885819212905816; … ; 
0.773543733306263 0.10880091279797399 … 0.10525512055751185 0.6303472234021711; 
0.10845217539341823 0.9911071602976902 … 0.014754069216096566 0.5256103389041187], 
[0.4302689295553531, 0.2396683446618325, 0.7954433314513768, 0.7191098533903124, 
0.8168563428651753, 0.7064320936729905, 0.048113106979693065, 0.3102938851371281, 
0.6246380539228858, 0.3510284321966193  …  0.5324022501182528, 0.8354720951777901, 
0.7526652774981095, 0.3639742621882005, 0.21030903031988923, 0.6936212944871928, 
0.3910592143534404, 0.15152013651215301, 0.38891692138831735, 0.08827711410802941], 
Float64[0, 0, 1, 1, 0, 1, 0, 0, 1, 0  …  0, 0, 1, 1, 1, 1, 0, 1, 0, 0])