ATE/ATT/ITT Estimation

Structs and methods to estimate the ATE, ATT, ITT, and change over time from observational research designs.

CausalELM.Estimators — Module

Estimate causal effects with interrupted time series analysis, G-computation, and double machine learning using Extreme Learning machines.

CausalELM.Estimators.CausalEstimator — Type

Abstract type for GComputation and DoubleMachineLearning

CausalELM.Estimators.InterruptedTimeSeries — Type

Container for the results of an interrupted time series analysis

CausalELM.Estimators.GComputation — Type

Container for the results of G-Computation

CausalELM.Estimators.DoubleMachineLearning — Type

Container for the results of a double machine learning estimator

CausalELM.estimate_causal_effect! — Function

estimate_causal_effect!(its)

Estimate the effect of an event relative to a predicted counterfactual.

Examples

julia> X₀, Y₀, X₁, Y₁ =  rand(100, 5), rand(100), rand(10, 5), rand(10)
julia> m1 = InterruptedTimeSeries(X₀, Y₀, X₁, Y₁)
julia> estimate_causal_effect!(m1)
0.25714308

estimate_causal_effect!(g)

Estimate a causal effect of interest using G-Computation.

If treatents are administered at multiple time periods, the effect will be estimated as the average difference between the outcome of being treated in all periods and being treated in no periods.For example, given that individuals 1, 2, ..., i ∈ I recieved either a treatment or a placebo in p different periods, the model would estimate the average treatment effect as E[Yᵢ|T₁=1, T₂=1, ... Tₚ=1, Xₚ] - E[Yᵢ|T₁=0, T₂=0, ... Tₚ=0, Xₚ].

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m2 = GComputation(X, Y, T)
julia> estimate_causal_effect!(m2)
0.31067439

estimate_causal_effect!(DML)

Estimate a causal effect of interest using double machine learning.

Unlike other estimators, this method does not support time series or panel data. This method also does not work as well with smaller datasets because it estimates separate outcome models for the treatment and control groups.

Examples

julia> X, Xₚ, Y, T =  rand(100, 5), rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m3 = DoubleMachineLearning(X, Xₚ, Y, T)
julia> estimate_causal_effect!(m3)
0.31067439

estimate_causal_effect!(m)

Estimate the CATE using a metalearner.

For an overview of meatlearning see:

Künzel, Sören R., Jasjeet S. Sekhon, Peter J. Bickel, and Bin Yu. "Metalearners for 
estimating heterogeneous treatment effects using machine learning." Proceedings of the 
national academy of sciences 116, no. 10 (2019): 4156-4165.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m4 = SLearner(X, Y, T)
julia> estimate_causal_effect!(m4)
[0.20729633391630697, 0.20729633391630697, 0.20729633391630692, 0.20729633391630697, 
0.20729633391630697, 0.20729633391630697, 0.20729633391630697, 0.20729633391630703, 
0.20729633391630697, 0.20729633391630697  …  0.20729633391630703, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630703, 0.20729633391630697, 0.20729633391630697, 
0.20729633391630692, 0.20729633391630697, 0.20729633391630697, 0.20729633391630697]

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m5 = TLearner(X, Y, T)
julia> estimatecausaleffect!(m5)
[0.0493951571746305, 0.049395157174630444, 0.0493951571746305, 0.049395157174630444, 
0.04939515717463039, 0.04939515717463039, 0.04939515717463039, 0.04939515717463039, 
0.049395157174630444, 0.04939515717463061  …  0.0493951571746305, 0.04939515717463039, 
0.0493951571746305, 0.04939515717463039, 0.0493951571746305, 0.04939515717463039, 
0.04939515717463039, 0.049395157174630444, 0.04939515717463039, 0.049395157174630444]

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = XLearner(X, Y, T)
julia> estimatecausaleffect!(m1)
[-0.025012644892878473, -0.024634294305967294, -0.022144246680543364, -0.023983138957276127, 
-0.024756239357838557, -0.019409519377053822, -0.02312807640357356, -0.016967113188439076, 
-0.020188871831409317, -0.02546526148141366  …  -0.019811641136866287, 
-0.020780821058711863, -0.013588359417922776, -0.020438648396328824, -0.016169487825519843, 
-0.024031422484491572, -0.01884713946778991, -0.021163590874553318, -0.014607310062509895, 
-0.022449034332142046]

CausalELM.Estimators.first_stage! — Function

first_stage!(DML, x₀, xₚ, T, y₀)

Estimate the average treatment for the treated for a double machine learning estimator with cross fitting.

Examples

julia> X, Xₚ Y, T =  rand(100, 5), rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = DoubleMachineLearning(X, Xₚ, Y, T)
julia> x₀, y₀ = m1.X[m1.T .== 0], m1.Y[m1.T .== 0]
julia> first_stage!(m1, x₀, xₚ, T, y₀)
(Regularized Extreme Learning Machine with 10 hidden neurons, 
Regularized Extreme Learning Machine with 10 hidden neurons)

CausalELM.Estimators.ate! — Function

ate!(DML, x₁, y₁)

Estimate the average treatment effect for a double machine learning estimator with cross fitting.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> m1 = DoubleMachineLearning(X, Y, T)
julia> x₁, y₁ = m1.X[m1.T .== 1], m1.Y[m1.T .== 1]
julia> ate!(m1, x₁, y₁)
Regularized Extreme Learning Machine with 10 hidden neurons

CausalELM.Estimators.predict_propensity_score — Function

predict_propensity_score(ps_model, x_pred)

Predict the propensity score for an out of sample fold for the double machine learning estimator.

Examples

julia> X, Xₚ, Y, T =  rand(100, 5), rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> x_pred = rand(20, 5)
julia> m1 = DoubleMachineLearning(X, Xₚ, Y, T)
julia> ps_model, _ = first_stage!(m1, x₀, y₀)
julia> predict_propensity_score(ps_model, x_pred)

CausalELM.Estimators.predict_control_outcomes — Function

predict_control_outcomes(control_model, x_pred)

Predict the counterfactual control outcomes for an out of sample fold.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> x_pred = rand(20, 5)
julia> m1 = DoubleMachineLearning(X, Y, T)
julia> control_model, _ = first_stage!(m1, x₀, y₀)
julia> predict_control_outcomes(control_model, x_pred)

CausalELM.Estimators.predict_treatment_outcomes — Function

predict_treatment_outcomes(treatment_model, x_pred)

Predict the counterfactual treatment outcomes for an out of sample fold.

Examples

julia> X, Y, T =  rand(100, 5), rand(100), [rand()<0.4 for i in 1:100]
julia> x_pred = rand(20, 5)
julia> m1 = DoubleMachineLearning(X, Y, T)
julia> treatment_model, _ = first_stage!(m1, x₀, y₀)
julia> predict_treatment_outcomes(treatment_model, x_pred)

CausalELM.Estimators.crossfitting_sets — Function

crossfitting_sets(DML)

Creates folds for cross fitting a double machine learning estimator.

Examples

julia> xfolds, y_folds = crossfiting_sets(DML)
([[0.0 0.0; 0.0 0.0; 0.0 0.0; 0.0 0.0], [0.0 0.0; 0.0 0.0; 0.0 0.0; 0.0 0.0], 
[0.0 0.0; 0.0 0.0; 0.0 0.0], [0.0 0.0; 0.0 0.0; 0.0 0.0; 0.0 0.0], [0.0 0.0; 0.0 0.0; … ; 
0.0 0.0; 0.0 0.0]], [[0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 
0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0]])

CausalELM.Estimators.moving_average — Function

moving_average(g)

Calculates a cumulative moving average.

Examples

julia> moving_average([1, 2, 3])
3-element Vector{Float64}
1.0
1.5
2.0